APC환경설정 상품등록, 품목별 규격/감량률 등록 수정

+++ src/main/java/com/at/apcss/am/cmns/vo/CmnsSpcfctVO.java
@@ -47,7 +47,7 @@ public class CmnsSpcfctVO extends CmnsItemVO {
 	/**
 	 * 단중
 	 */
-	private float wght;
+	private double wght;
 
 	/**
 	 * 순번
